色々リファクタした

+++ src/main/java/com/makkii/maic/training/EmbeddingTrainer.java
@@ -1,10 +1,10 @@
 package com.makkii.maic.training;
 
-import java.io.*;
+import java.io.IOException;
+import java.io.ObjectInputStream;
+import java.io.ObjectOutputStream;
 import java.nio.file.Files;
 import java.nio.file.Paths;
-import java.util.Dictionary;
-import java.util.Random;
 
 import static com.makkii.maic.AIParams.*;
 
@@ -13,17 +13,17 @@ public class EmbeddingTrainer {
     /**
      * 埋め込み行列を学習(更新)します。
      *
-     * @param inputIndices  バッチ内の単語のインデックスの配列
-     * @param gradients     対応する勾配の配列 (embeddingMatrix と同じ次元)
-     * @param learningRate  学習率
+     * @param inputIndices バッチ内の単語のインデックスの配列
+     * @param gradients    対応する勾配の配列 (embeddingMatrix と同じ次元)
+     * @param learningRate 学習率
      */
-    public static void train(int[] inputIndices, float[][] gradients, float learningRate) {
+    public static void train(int[] inputIndices, double[][] gradients, double learningRate) {
         for (int i = 0; i < inputIndices.length; i++) {
             int wordIndex = inputIndices[i]; // 更新対象の単語のインデックス
 
             // inputIndicesのサイズとgradientsのサイズは同じため、
             // inputIndices.length == gradients.length になる
-            float[] gradient = gradients[i];  // 対応する勾配
+            double[] gradient = gradients[i];  // 対応する勾配
 
             // 勾配降下法による更新
             for (int j = 0; j < DIMENSION; j++) {
@@ -36,9 +36,9 @@ public class EmbeddingTrainer {
      * 指定された単語インデックスの埋め込みベクトルを取得します.
      *
      * @param index 単語のインデックス
-     * @return 埋め込みベクトル (float配列)
+     * @return 埋め込みベクトル (double配列)
      */
-    public static float[] getEmbedding(int index) {
+    public static double[] getEmbedding(int index) {
         // 境界チェック(本番環境では例外処理などが推奨)
         if (index < 0 || index >= wordSize) {
             throw new IllegalArgumentException("Invalid word index: " + index);
@@ -67,15 +67,16 @@ public class EmbeddingTrainer {
      */
     public void load(String filePath) throws IOException, ClassNotFoundException {
         try (ObjectInputStream ois = new ObjectInputStream(Files.newInputStream(Paths.get(filePath)))) {
-            embeddingMatrix = (float[][]) ois.readObject();
+            embeddingMatrix = (double[][]) ois.readObject();
         }
     }
 
     /**
      * 既存の埋め込み行列を設定します
+     *
      * @param matrix
      */
-    public void setEmbeddingMatrix(float[][] matrix) {
+    public void setEmbeddingMatrix(double[][] matrix) {
         if (matrix.length != wordSize || matrix[0].length != DIMENSION) {
             throw new IllegalArgumentException("matrix size is not match");
         }
@@ -87,19 +88,19 @@ public class EmbeddingTrainer {
 
         // ダミーデータの作成(本来は学習データから)
         int[] inputIndices = {0, 1, 2};
-        float[][] gradients = {
+        double[][] gradients = {
                 {0.1f, -0.2f, 0.05f, /*...*/}, // dimensionの数だけ要素が必要
                 {-0.03f, 0.1f, -0.1f, /*...*/},
                 {0.0f, 0.01f, -0.02f, /*...*/}
         };
-        float learningRate = 0.01f;
+        double learningRate = 0.01f;
 
         // 学習の実行例
         train(inputIndices, gradients, learningRate);
 
         //埋め込み行列の一部を出力
-        float[] embedding = getEmbedding(0);
-        for(int i=0; i<10; ++i) {
+        double[] embedding = getEmbedding(0);
+        for (int i = 0; i < 10; ++i) {
             System.out.print(embedding[i] + " ");
         }
     }
